Move transpiler settings to respective packages
Finish making main() arg parsing dynamic.
