+++ tests/test_transpile_self.py
@@ -160,8 +160,8 @@ class SelfTranspileTests(unittest.TestCase):
                 "annotation_transformer.py",
                 "ast_helpers.py",
                 "astx.py",
-                "context.py",
                 "cli.py",
+                "context.py",
                 "declaration_extractor.py",
                 "exceptions.py",
                 "language.py",
@@ -261,5 +261,5 @@ class SelfTranspileTests(unittest.TestCase):
                 _suppress_exceptions=(AstNotImplementedError, AstUnrecognisedBinOp),
             ),
             format_error_count=2,
-            failure_count=12,
+            failure_count=13,
         )
